+++ Stage_OnAir/Assets/02_Ingame/Scripts/Manager/GoogleAdsManager.cs
@@ -138,7 +138,7 @@ public class GoogleAdsManager : Singleton<GoogleAdsManager>
 
     private void userEarnedRewardCallback(Reward reward)
     {
-        int Result = Mathf.CeilToInt(GameManager.Instance.Money * 0.1f);
+        long Result = Mathf.CeilToInt(GameManager.Instance.Money * 0.1f);
         GameManager.Instance.CostMoney(Result, false);
     }
     private void HandleAdFailedToPresent(object sender, AdErrorEventArgs args)
@@ -200,8 +200,8 @@ public class GoogleAdsManager : Singleton<GoogleAdsManager>
         string type = args.Type;
         double amount = args.Amount;
 
-        int Money = Mathf.CeilToInt(GameManager.Instance.Money * 0.1f);
-        int Result = Mathf.Abs(Money);
+        long Money = Mathf.CeilToInt(GameManager.Instance.Money * 0.1f);
+        long Result = (long)Mathf.Abs(Money);
         GameManager.Instance.CostMoney(Result, false);
 
         MonoBehaviour.print(
